Вынес DTO в отдельный класс
