+++ Saber/Java/src/main/java/ru/beetlerat/saber/lists/ListStringSerialize.java
@@ -8,39 +8,6 @@ public class ListStringSerialize extends ListRand {
     private static final int NO_LINK = -1, BROKEN_LINK = -2;
     NodeDTO[] nodesDTO;
 
-    public class NodeDTO {
-        public int ID; // Добавлено для поддержания ссылочной целостности при строковой реализации
-        public int PrevID;
-        public int NextID;
-        public int RandID; // произвольный элемент внутри списка
-        public String Data;
-
-        private ListNode originalNode;
-
-        public NodeDTO(int ID, String data, int prevID, int nextID, int randID) {
-            this.ID = ID;
-            this.PrevID = prevID;
-            this.NextID = nextID;
-            this.RandID = randID;
-            this.Data = data;
-        }
-
-
-        public String toString() {
-            StringBuilder json = new StringBuilder();
-
-            json.append("{\n");
-            json.append("   \"ID\": ").append(ID).append("\n");
-            json.append("   \"Data\": ").append(Data).append("\n");
-            json.append("   \"PreviousID\": ").append(PrevID).append("\n");
-            json.append("   \"NextID\": ").append(NextID).append("\n");
-            json.append("   \"RandID\": ").append(RandID).append("\n");
-            json.append("}\n");
-
-            return json.toString();
-        }
-    }
-
     @Override
     public void Serialize(FileOutputStream s) {
         nodesDTO = new NodeDTO[Count];
